+++ src/frontend/tests/starter-projects/Vector Store.spec.ts
